Bump version to v2.0.8
- Updated version in `kterm.h` and `README.md`.
- Added changelog entry for v2.0.8 in `kterm.h` highlighting:
  - Situation Decoupling via Aliasing (Phase 2).
  - Restoration of Printer Controller and ReGIS functionality.
  - Fixes for hardcoded screen dimensions.
  - Architecture updates using `KTerm*` types.

+++ tests/test_vttest_suite.c
@@ -15,7 +15,7 @@ void VerifyScreenCell(int row, int col, char expected_char, int fg_idx, int bg_i
         return;
     }
 
-    if (cell->ch != expected_char) {
+    if (cell->ch != (unsigned int)expected_char) {
         printf("FAIL at %d,%d: Expected char '%c', got '%c'\n", row, col, expected_char, (char)cell->ch);
     }
